requirements.txt
pynput
tk
pywin32
pystray
Pillow
